feat(notes): WP6 -- /notes/:id note page + QuickNoteComposer capture unification
NotePage (new note-page manifest surface, public tier, mirrors the
/collections/:id pattern): deep-linkable full-page editor with back nav,
reading/split/rail toggles, single header title, autosave, properties rail,
backlinks, and an outline rail for 3+-heading notes (fence-aware parser;
click scrolls the editor via MarkdownEditorHandle.scrollToLine or the nth
rendered heading in reading view). Notes list cards/rows navigate here.

QuickNoteComposer is now the single creation surface: Notes "New", gap
preloads, reader selection capture, and reader section notes (autosave
createExtras stamps document_id/section_id; CreateNotePayload gained
section_id). Keeps append-to-existing; adds an "Open full note" hatch that
flushes then navigates.

Deleted: NoteReaderSheet.tsx (zero consumers) and SectionListItem's bespoke
inline note editor. Also rides along: dead group filter removed from
Notes.tsx and the list-view title cell now shows note.title.

Live-drive verified: card->page URL, deep-link fresh load, back nav, outline
render + scroll, composer -> open-full-note -> delete round trip, section
capture POST carries section_id.
